requirements.txt
# TrustPoints Backend Dependencies
# Python 3.12 compatible

# Flask Framework
Flask==3.0.0
Werkzeug==3.0.1

# MongoDB
Flask-PyMongo==2.3.0
pymongo==4.6.1

# JWT Authentication
PyJWT==2.8.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# CORS Support (for frontend integration)
Flask-CORS==4.0.0

# Production Server
gunicorn==21.2.0
